Drawing pieces.
Added new functions to draw the pieces in the board. Also added a
function to handle the mouse pressed event.

In this commit I changed the old index ranges from [1,3] to [0,2],
because there's no need to keep it in the old form. I hope I made all
the changes so I don't break anything.

I also changed the TICKS_WIDTH variable to draw the pieces correctly.